requirements.txt
asttokens==2.4.1
autopep8==2.3.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
descartes==1.1.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fire==0.6.0
fonttools==4.53.0
fsspec==2024.2.0
gguf==0.9.1
huggingface-hub==0.24.0
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter-client==7.3.4
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nuscenes-devkit==1.1.11
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.8
pycodestyle==2.12.0
Pygments==2.18.0
pyparsing==3.1.2
pyquaternion==0.9.9
pytest==8.2.2
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==69.5.1
Shapely==1.8.5.post1
six==1.16.0
stack-data==0.6.2
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
ipykernel==6.29.4
ipython==8.25.0
tornado==6.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.43.0
waymo_open_dataset_tf_2_12_0==1.6.2